requirements.txt
APScheduler==3.6.0
diskcache==5.1.0
requests
websockets>=7.0.0
cryptography==2.5
python-json-logger
pytz==2019.1
tensorflow==1.5.0
python-dateutil
Pillow
flask>=0.12.0
jinja2
imutils
psutil
netifaces
statsd
wget==3.2
pyyaml>=3.0
opencv-python==3.4.3.18
zeroconf==0.21.3
pyserial==3.4
future==0.18.1
gphoto2==2.1.0
-e git+https://github.com/turingvideo/python-common@0.0.4#egg=turingvideo
speedtest-cli==2.1.2
